+++ src/components/Layout/layout.module.scss
@@ -12,7 +12,7 @@
   }
 
   svg {
-    fill: currentColor;
+    fill: currentcolor;
   }
 }
