+++ src/routes/Dashboard/ChartHeader/SelectButton/Dropdown.tsx
@@ -3,7 +3,7 @@ import { useClickAway } from 'react-use'
 import { useSetRecoilState } from 'recoil'
 import cx from 'classnames'
 
-import { selectorState } from '../../states/dashBoard'
+import { selectorState } from 'routes/Dashboard/states/dashBoard'
 import { Circle } from 'assets/svgs'
 import styles from './selectButton.module.scss'
