+++ src/routes/Dashboard/ChartHeader/SelectButton/index.tsx
@@ -2,8 +2,9 @@ import { useState, MouseEvent } from 'react'
 import { useRecoilValue } from 'recoil'
 import cx from 'classnames'
 
-import { selectorState } from '../../states/dashBoard'
+import { selectorState } from 'routes/Dashboard/states/dashBoard'
 import Dropdown from './Dropdown'
+
 import { Circle, ArrowDownIcon } from 'assets/svgs'
 import styles from './selectButton.module.scss'
